requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
blis==1.3.3
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   nltk
    #   typer
    #   typer-slim
cloudpathlib==0.23.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cymem==2.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
docopt==0.6.2
    # via num2words
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
    # via -r requirements.in
fonttools==4.61.1
    # via matplotlib
idna==3.11
    # via requests
jinja2==3.1.6
    # via spacy
joblib==1.5.3
    # via nltk
kiwisolver==1.4.9
    # via matplotlib
langcodes==3.5.1
    # via spacy
language-tool-python==2.9.4
    # via -r requirements.in
lxml==6.0.2
    # via python-docx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
matplotlib==3.9.4
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.15
    # via
    #   preshed
    #   spacy
    #   thinc
nltk==3.9.1
    # via -r requirements.in
num2words==0.5.14
    # via -r requirements.in
numpy==2.4.0
    # via
    #   blis
    #   contourpy
    #   matplotlib
    #   pandas
    #   spacy
    #   thinc
packaging==25.0
    # via
    #   matplotlib
    #   spacy
    #   thinc
    #   weasel
pandas==2.3.0
    # via -r requirements.in
pillow==12.0.0
    # via matplotlib
preshed==3.0.12
    # via
    #   spacy
    #   thinc
psutil==7.0.0
    # via
    #   -r requirements.in
    #   language-tool-python
pydantic==2.12.5
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-docx==1.2.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via -r requirements.in
rapidfuzz==3.13.0
    # via -r requirements.in
regex==2025.11.3
    # via nltk
requests==2.32.4
    # via
    #   -r requirements.in
    #   language-tool-python
    #   spacy
    #   weasel
rich==14.0.0
    # via
    #   -r requirements.in
    #   typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.5.0
    # via weasel
spacy==3.8.7
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
thinc==8.3.10
    # via spacy
toml==0.10.2
    # via language-tool-python
tqdm==4.67.1
    # via
    #   language-tool-python
    #   nltk
    #   spacy
typer==0.20.1
    # via spacy
typer-slim==0.20.1
    # via weasel
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.3
    # via spacy
wrapt==2.0.1
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
